+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Removed dependency on `purrr`. 
 * Removed dependency on `tibble`.
+* Removed dependency on `scda` and `scda.2022` and used `mmrm_test_data` as the sample dataset for tests and example apps.
 
 # tern.mmrm 0.1.0